requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
paddleocr>=3.0.0
paddlepaddle>=3.0.0
python-multipart>=0.0.9
pydantic>=2.0.0
pydantic-settings>=2.0.0
aiofiles>=24.0.0
Pillow>=10.0.0
opencv-python-headless>=4.8.0
numpy>=1.24.0
python-dotenv>=1.0.0
